+++ src/include/ecoz2.h
@@ -32,6 +32,34 @@ int ecoz2_prd_show_file(
         int to
         );
 
+int ecoz2_vq_learn(
+        int prediction_order,
+        double epsilon,
+        const char *codebook_class_name,
+        const char *predictor_filenames[],
+        int num_predictors
+        );
+
+int ecoz2_vq_quantize(
+        const char *nom_raas,
+        const char *predictor_filenames[],
+        int num_predictors
+        );
+
+int ecoz2_vq_show(
+        char *codebook_filename,
+        int from,
+        int to
+        );
+
+int ecoz2_vq_classify(
+        char **cb_filenames,
+        int num_codebooks,
+        char **prd_filenames,
+        int num_predictors,
+        int show_ranked
+        );
+
 // double to facilitate rust wrapper
 typedef void (*hmm_learn_callback_t)(char*, double);
 
@@ -44,6 +72,28 @@ int ecoz2_hmm_learn(
         double val_auto,
         int max_iterations,
         hmm_learn_callback_t callback
+        );
+
+int ecoz2_hmm_classify(
+        char **model_names,
+        int num_model_names,
+        char **seq_filenames,
+        int num_seq_filenames,
+        int show_ranked
+        );
+
+int ecoz2_hmm_show(
+        char *filename,
+        char *format
+        );
+
+int ecoz2_seq_show_files(
+        int with_prob,
+        int gen_Qopt,
+        int no_sequence,
+        char* hmm_filename,
+        char* seq_filenames[],
+        int num_seq_filenames
         );
 
 #ifdef __cplusplus